fix: fetch full view for schema diagram
